New and improved
In the controller, improved save and get report methods. In the save method, it checks if there is already a file created for the patient. If it evaluates that there is no file it will then create a file for the user. Pressing save a second time will then allow the save properly. I also update the save and get report methods to only look and create files in the C: root drive. Looking for, creating, and saving files in the C: root will allow compatibility across any Windows machine.