Description
In this project, I am testing the creation of Java Threads. In the main file, there is only the instantiation of the Simulation object that then called the runSim() method. The runSim() method then uses the Executor Service to create three concurrent threads, as well as, a synchronized print method for the threads. In the Generator thread, an ArrayList is filled with 10 random integers ranging from 1-10, after, it will print the ArrayList's current state, lastly, it passes control to the Sorter thread. The sorter thread checks to see if ArrayList is full or not, then it will sort the ArrayList and print it.  Last is the Drainer thread, if the Sorter has finished the drainer will then clear the ArrayList and print it. Inside all of the overriden run methods for the threads is a for loop that will execute each thread 10 times.